staticanalysis/bot: Move mercurial output process to cli_common
